requirements_dev.txt
appdirs==1.4.4
attrs==21.2.0
beautifulsoup4==4.9.3
black==21.5b1
bs4==0.0.1
certifi==2020.12.5
cfgv==3.2.0
chardet==4.0.0
click==8.0.0
colorama==0.4.4
commonmark==0.9.1
coverage==5.5
distlib==0.3.1
filelock==3.0.12
flake8==3.9.2
identify==2.2.4
idna==2.10
iniconfig==1.1.1
mccabe==0.6.1
mutagen==1.45.1
mypy==0.812
mypy-extensions==0.4.3
nodeenv==1.6.0
packaging==20.9
pathspec==0.8.1
pluggy==0.13.1
pre-commit==2.12.1
pre-commit-hooks==4.0.0
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.9.0
pyparsing==2.4.7
pytest==6.2.4
pytest-cov==2.12.0
pytest-mock==3.6.1
python-slugify==5.0.2
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
rich==10.2.2
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
six==1.16.0
soupsieve==2.2.1
text-unidecode==1.3
toml==0.10.2
tox==3.23.1
typed-ast==1.4.3
typing-extensions==3.10.0.0
urllib3==1.26.4
virtualenv==20.4.6
